+++ packages/orq-rc/RELEASES.md
@@ -2539,3 +2539,13 @@ Based on:
 - [python v3.12.0-rc.30] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.30] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.30 - packages/orq-rc
+
+## 2025-09-03 10:18:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.607.1 (2.690.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.31] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.31] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.31 - packages/orq-rc